Moved form into components folder
