+++ frontend/app/register/page.tsx
@@ -1,4 +1,4 @@
-import Form from "@/app/register/form";
+import Form from "@/app/(components)/registerForm";
 
 export default async function RegisterPage() {
